fix: split monolithic scripts for issue #2515 (syngas + controls_tab)
Split syngas_compression_calculator.py (1171 LOC) and controls_tab.py
(1131 LOC) into focused modules, bringing each to <= 300 LOC per the
issue acceptance criteria.

Syngas split:
- syngas_compression_engine.py: CompressionStage + SyngasCompressionEngine
  (pure Python, no Qt dependency)
- syngas_compression_worker.py: CompressionCalculationWorker thread
- syngas_compression_tabs_mixin.py: tab builder mixin (_SyngasTabsMixin)
- syngas_compression_display.py: standalone format/render functions
- syngas_compression_calculator.py: trimmed to 276 LOC (widget + factory)

Controls tab split:
- actuator_detail_dialog.py: ActuatorDetailDialog dialog class
- actuator_controls_mixin.py: _ActuatorControlsMixin (actuator management)
- kinematic_controls_mixin.py: _KinematicControlsMixin (joint sliders)
- simulation_controls_mixin.py: _SimulationControlsMixin (playback handlers)
- controls_tab.py: trimmed to 280 LOC (core setup + signal handlers)

Tests added:
- tests/unit/test_script_size_contracts_2515.py: LOC budget assertions
- tests/unit/test_syngas_compression_engine.py: engine + display tests

Closes #2515
